requirements.txt
livekit-agents[openai]
python-dotenv